requirements.txt
numpy
opencv-python
flask
tensorflow
livereload
pillow